// src/app/model/config/orden.model.ts
export class Orden {
    id: number;
    idTipoOrden: number;
    nomTipoOrden: string;

    nroOrden: string;
    nroExpSIAF: string;
    fecha: Date;
    idTipoDocumento: number;
    nomTipoDocumento: string;
    nroDocumento: string;
    idProveedor: number;
    nomProveedor: string;
    monto: number;
    idEstado: number;
    nomEstado: string;
}